removed old samples
