--- sdk/storage/storage-queue/samples/v12/javascript/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@azure-samples/storage-queue-js",
-  "private": true,
-  "version": "1.0.0",
-  "description": "Azure Storage Queue client library samples for JavaScript",
-  "engines": {
-    "node": ">=12.0.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
-    "directory": "sdk/storage/storage-queue"
-  },
-  "keywords": [
-    "azure",
-    "storage",
-    "queue",
-    "cloud",
-    "node.js",
-    "typescript",
-    "javascript",
-    "browser"
-  ],
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
-  },
-  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/storage/storage-queue",
-  "dependencies": {
-    "@azure/storage-queue": "latest",
-    "dotenv": "latest",
-    "@azure/identity": "^2.0.1"
-  }
-}
