--- sdk/storage/storage-queue/samples/v12/javascript/sample.env
@@ -1,20 +1,0 @@
-# Used in most samples. Retrieve these values from a storage account in the Azure Portal.
-ACCOUNT_NAME=<storage-account-name>
-ACCOUNT_KEY=<storage-account-key>
-
-# Used for withConnString
-STORAGE_CONNECTION_STRING=<storage-account-connection-string>
-
-# Used for the advanced and anonymousCred tests. Create a SAS token for a storage account in the Azure Portal.
-ACCOUNT_SAS=<shared-access-signature>
-
-# Used to authenticate using Azure AD as a service principal for role-based authentication.
-#
-# See the documentation for `EnvironmentCredential` at the following link:
-# https://docs.microsoft.com/javascript/api/@azure/identity/environmentcredential
-AZURE_TENANT_ID=<AD tenant id or name>
-AZURE_CLIENT_ID=<ID of the user/service principal to authenticate as>
-AZURE_CLIENT_SECRET=<client secret used to authenticate to Azure AD>
-
-# To run the proxyAuth sample, set up an HTTP proxy and enter your information:
-# HTTP_PROXY=http://localhost:3128
