--- sdk/storage/storage-queue/samples/v12/typescript/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "@azure-samples/storage-queue-ts",
-  "private": true,
-  "version": "1.0.0",
-  "description": "Azure Storage Queue client library samples for TypeScript",
-  "engines": {
-    "node": ">=12.0.0"
-  },
-  "scripts": {
-    "build": "tsc",
-    "prebuild": "rimraf dist/"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
-    "directory": "sdk/storage/storage-queue"
-  },
-  "keywords": [
-    "azure",
-    "storage",
-    "queue",
-    "cloud",
-    "node.js",
-    "typescript",
-    "javascript",
-    "browser"
-  ],
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
-  },
-  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/storage/storage-queue",
-  "dependencies": {
-    "@azure/storage-queue": "latest",
-    "dotenv": "latest",
-    "@azure/identity": "^2.0.1"
-  },
-  "devDependencies": {
-    "@types/node": "^12.0.0",
-    "typescript": "~4.4.0",
-    "rimraf": "latest"
-  }
-}
