--- sdk/storage/storage-queue/samples/v12/typescript/src/anonymousAuth.ts
@@ -1,46 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-/**
- * @summary authenticate anonymously using a SAS-encoded URL
- */
-
-import { QueueServiceClient, AnonymousCredential } from "@azure/storage-queue";
-
-// Load the .env file if it exists
-import * as dotenv from "dotenv";
-dotenv.config();
-
-export async function main() {
-  // Enter your storage account name and SAS
-  const account = process.env.ACCOUNT_NAME || "";
-  const accountSas = process.env.ACCOUNT_SAS || "";
-
-  // Use AnonymousCredential when url already includes a SAS signature
-  const anonymousCredential = new AnonymousCredential();
-
-  const queueServiceClient = new QueueServiceClient(
-    // When using AnonymousCredential, following url should include a valid SAS or support public access
-    `https://${account}.queue.core.windows.net${accountSas}`,
-    anonymousCredential
-  );
-
-  // Create a new queue
-  const queueName = `newqueue${new Date().getTime()}`;
-  const queueClient = queueServiceClient.getQueueClient(queueName);
-  const createQueueResponse = await queueClient.create();
-  console.log(
-    `Created queue ${queueClient.name} successfully, service assigned request ID: ${createQueueResponse.requestId}`
-  );
-
-  // Delete the queue.
-  const deleteQueueResponse = await queueClient.delete();
-  console.log(
-    `Deleted queue ${queueClient.name} successfully, service assigned request ID: ${deleteQueueResponse.requestId}`
-  );
-}
-
-main().catch((error) => {
-  console.error(error);
-  process.exit(1);
-});
